server: unify traits across protocol versions
