+++ rustls/src/lib.rs
@@ -633,8 +633,9 @@ pub mod server {
         };
     }
 
-    pub(crate) use tls12::{TLS12_HANDLER, Tls12Handler};
-    pub(crate) use tls13::{TLS13_HANDLER, Tls13Handler};
+    pub(crate) use hs::ServerHandler;
+    pub(crate) use tls12::TLS12_HANDLER;
+    pub(crate) use tls13::TLS13_HANDLER;
 }
 
 pub use server::ServerConfig;
