+++ rustls/src/versions.rs
@@ -1,4 +1,6 @@
 use crate::enums::ProtocolVersion;
+use crate::tls12::Tls12CipherSuite;
+use crate::tls13::Tls13CipherSuite;
 
 /// A TLS protocol version supported by rustls.
 ///
@@ -78,7 +80,7 @@ pub static TLS13_VERSION: &Tls13Version = &Tls13Version {
 #[derive(Debug)]
 pub struct Tls12Version {
     pub(crate) client: &'static dyn crate::client::Tls12Handler,
-    pub(crate) server: &'static dyn crate::server::Tls12Handler,
+    pub(crate) server: &'static dyn crate::server::ServerHandler<Tls12CipherSuite>,
 }
 
 /// Internal data for handling the TLS1.3 protocol.
@@ -88,5 +90,5 @@ pub struct Tls12Version {
 #[derive(Debug)]
 pub struct Tls13Version {
     pub(crate) client: &'static dyn crate::client::Tls13Handler,
-    pub(crate) server: &'static dyn crate::server::Tls13Handler,
+    pub(crate) server: &'static dyn crate::server::ServerHandler<Tls13CipherSuite>,
 }
